feat: handle shared playlists in back/front and various rights in back

Closes #55

See merge request pwa-group-b/doremix!80

+++ src/back/controllers/playlist.py
@@ -16,6 +16,10 @@ class PlaylistController:
         return PlaylistRepository.get_public_playlists(db)
 
     @staticmethod
+    def get_shared_playlists(db: Session, user_id: int):
+        return PlaylistRepository.get_shared_playlist(db, user_id)
+
+    @staticmethod
     def get_playlist(db: Session, playlist_id: int, user_id: int):
         playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
         if not playlist:
@@ -79,36 +83,40 @@ class PlaylistController:
         return PlaylistRepository.create(db, new_playlist)
 
     @staticmethod
-    def delete_playlist(db: Session, playlist_id: int, user_id: int):
-        # TODO: Quand l'auth sera en place, ajouter user_id en paramètre :
-        # def delete_playlist(db: Session, playlist_id: int, user_id: int):
-
-        playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
+    def delete_playlist(db: Session, playlist_id: int, user: User):
+        playlist = PlaylistRepository.get_by_id(db, playlist_id, user.idUser)
 
         if not playlist:
             raise HTTPException(status_code=404, detail="Playlist not found")
 
-        # TODO: Quand l'auth sera en place, vérifier que l'utilisateur est le propriétaire :
-        # if playlist.idOwner != user_id:
-        #     raise HTTPException(status_code=403, detail="You are not the owner of this playlist")
+        # Only playlist owner and admin can delete playlist
+        if not (playlist.idOwner == user.idUser or user.idRole == 3):
+            raise HTTPException(
+                status_code=404, detail="You're not allowed to delete this playlist."
+            )
 
         PlaylistRepository.delete(db, playlist)
 
         return {"message": f"Playlist '{playlist.name}' successfully deleted"}
 
     @staticmethod
-    def update_playlist(db: Session, playlist_id: int, update_data: dict, user_id: int):
-        # TODO: Quand l'auth sera en place, ajouter user_id en paramètre :
-        # def update_playlist(db: Session, playlist_id: int, update_data: dict, user_id: int):
-
-        playlist = PlaylistRepository.get_by_id(db, playlist_id, user_id)
+    def update_playlist(db: Session, playlist_id: int, update_data: dict, user: User):
+        playlist = PlaylistRepository.get_by_id(db, playlist_id, user.idUser)
+        users, _ = PlaylistRepository.list_shared_user(db, playlist_id, user.idUser)
+        editors = [user.idUser for user in users if user.editor]
 
         if not playlist:
             raise HTTPException(status_code=404, detail="Playlist not found")
 
-        # TODO: Quand l'auth sera en place, vérifier que l'utilisateur est le propriétaire :
-        # if playlist.idOwner != user_id:
-        #     raise HTTPException(status_code=403, detail="You are not the owner of this playlist")
+        # Only playlist owner, playlist editors and admin can update playlist
+        if not (
+            playlist.idOwner == user.idUser
+            or user.idRole == 3
+            or user.idUser in editors
+        ):
+            raise HTTPException(
+                status_code=404, detail="You're not allowed to update this playlist."
+            )
 
         return PlaylistRepository.update_playlist(db, playlist, update_data)
 
@@ -134,6 +142,15 @@ class PlaylistController:
             raise HTTPException(500, "Failed to add track")
 
         return track
+
+    @staticmethod
+    def shared_with(db: Session, playlist_id: int, current_user_id: int):
+        users, err = PlaylistRepository.list_shared_user(
+            db, playlist_id, current_user_id
+        )
+        if err:
+            raise HTTPException(403, err)
+        return users
 
     @staticmethod
     def share_user(
